DA-15: Improve timeline view and add visualization config in scheme
The width calculation in the timeline is now dynamically calculated depending
on the max width of the neighbour nodes. Subgraphs are displayed in their own
section. Additionally the visualization elements can be configured in the scheme.
The kind if visualization (timeline and graph) can be configured also in the
scheme builder.