DA-15: Allow visualization configuration in scheme of none, graph and timeline
The scheme allows to define different visualizations of none, graph and timeline.
Furthermore display options were added for e.g. positioning the links in the
timeline mode. Therefore the scheme builder and the scheme preview was adjusted.
The entity models were changed so that every label and linklabel are now identified
by there id and not by there name/ string label, so that configurations to each
label can be applied.
The timeline places the nodes depending on the given configuration.
The visualization options are no exclusive.
Some bugs were fixed according to the displaying of the visualization kinds
schemeUploadModal html and corresponding controller has been refactored.