DA-199: Fix annotation selection by hotkey
Previously, selecting an annotation with the arrow keys would not
take the previous selection into account if that selection
had been made via the mouse or the graph. If an annotation outside
of the viewport was selected via the arrow keys, the view
was also not automatically adjusted.
Both of these issues have been fixed to allow for more intuitive
document navigation.