DA-200: Optimize removing projects
Removing a project which is related to removing documents and annotations etc.
was very inefficient. It is still not very fast with a lot of dependencies but
better. Sometimes a OutOfMemory exception occured. This should not happen
anymore evoked by deleting a project.
Instead of SELECT queries and then deleting the entites the queries were
replaced by direct DELETE queries.