DA-231: Fix rendering of links to/from off-screen annotations
Previously, links which had an off-screen annotations on one end were not
displayed correctly (either in the wrong direction or not at all). This is
fixed by estimating the positions of unrendered annotations by the position
of their first word so that the line can be drawn correctly.

NOTE: This patch fixes issues #231 and #232, as they are closely related.