DA-4+DA-70: Change persistence layer to prepare for future features
The persistence layer and database structure was changed, so a
feature of defining a colors for a scheme can be added in the future:
- to each scheme a colorScheme is now assigned, which can store
  colors for colorable entities
- for now, the colorMode of a colorScheme is always set to 'automatic'
  and no colors are assigned as the feature is not yet implemented
- spanTypes, labels, labelSets, linkLabels and linkTypes now inherit
  from a class ColorableBaseEntity, which extends BaseEntity.
  As a result, SpanTypes have id's as primary keys now.

Additionally, the following fields were added to prepare for future
implementation of features:
- "labelMenuStyle" for labelSets and linkTypes, so in the future users can
  select a "dropdown" option for labelSets/linkTypes with many labels.
  Right now this is always set to "list" by default as the feature
  is not yet implemented.
- "undirected" for linkTypes, so in the future undirected links can be
  defined. Right now this is always set to "false" by default as the feature
  is not yet implemented.

The mapping between Labels and LabelSets / LinkLabels and LinkTypes
was changed from @ManyToMany to @ManyToOne, as one Label belongs to
exactly one LabelSet/LinkType.
The classes LabelLabelSetMap and LinkLabelLinkTypeMap are not
needed anymore and have been removed.