HotFix: Fix target upload
Recently the primary key for the spantype was changed to id, as a result
the target upload did not work because in the target file only the name
is specified and not the id. So the check and fetching the spantype had
to be changed.
Some error handlings were improved.