DA-294: Fix missing appliesToSpanTypes attribute
In certain circumstances, the appliesToSpanTypes attribute was empty for Schemes,
even though it had been defined and was present in the database. This resulted
in an inability to choose labels during annotation.
The problem has been fixed by changing the FetchType for the labelSets and spanTypes
to EAGER, which should make sure that all necessary information is fetched in time.