DA-289: Refactor and rework handling of line breaks
In order to fix several bugs and undesirable effects, this patch reworks large
parts of how line breaks are handled in the annotation view. Most importantly,
a new system for calculating the width of lines was added, so that the positions
of line breaks can now be determined with greater accuracy instead of relying on
coarse approximations of the average character width. Alongside the more general
refactoring, this has the following effects:
- The application does not crash anymore when the length of a word exceeds the
  length of the annotation box.
- Lines containing many characters of above-average length (e.g. in Chinese texts)
  are not cut off anymore.
- Line breaks are (in general) closer to the edge of the annotation box, resulting
  in a more economic use of horizontal space.
- The new system can be manually optimized for other languages and scripts as well.