DA-304: Fix issues with completion checkbox and make it visible for admins
Previously, the completion status was not displayed correctly in the
project navigation bar, and the completion checkbox was never checked when
the project was not opened from the project explorer.
The completion checkbox is now also visible and accessible for admins and
project managers, allowing them to change the completion state of a document.

Additionally, an error that sometimes occurred in d3Annotation.js when
checking the tokenization language due to an asynchronous request was fixed.